Implementing a new segregated network model
catraia now relies on a new component, catraia-net, that runs on it's own
network namespace in order to not pollute default namespace with bridges and veth
interfaces. catraia-net is responsible for setting up bridge nd container's
network and for reverse-proxying requests to containers.

All communication between catraia and catraia-net occurs through unix sockets.
catraia-net listen to two sockets events.sock and proxy.sock. When a container
task is created, a event is sent through event.sock so catraia-net can setup
the network to the new container. Requests directed to containers, port 2020,
are tunneled to catraia-net through proxy.sock. catraia-net then proxys the
request to the apropriate container. While cumbersome this network model attends
to catraia's needs.